Show a confirmation dialog when deleting a scan Show a trash bin icon when swiping the tiles. Change tiles' leading icons. Copy scan text to clipboard on long press.

+++ lib/widgets/list_tiles.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
 
 import '../providers/scan_list_provider.dart';
@@ -23,20 +24,72 @@ class ListTiles extends StatelessWidget {
         itemCount: scans.length,
         itemBuilder: (context, i) {
           return Dismissible(
-            key: UniqueKey(),
-            background: Container(color: Colors.red),
+            key: ValueKey(scans[i].id),
+            background: Container(
+              color: Colors.red,
+              child: Align(
+                alignment: Alignment.centerLeft,
+                child: Padding(
+                  padding: const EdgeInsets.only(left: 16.0),
+                  child: Icon(
+                    Icons.delete,
+                    color: Colors.white,
+                  ),
+                ),
+              ),
+            ),
+            secondaryBackground: Container(
+              color: Colors.red,
+              child: Align(
+                alignment: Alignment.centerRight,
+                child: Padding(
+                  padding: const EdgeInsets.only(right: 16.0),
+                  child: Icon(
+                    Icons.delete,
+                    color: Colors.white,
+                  ),
+                ),
+              ),
+            ),
             onDismissed: (_) {
-              Provider.of<ScanListProvider>(context, listen: false)
-                  .deleteScanById(scans[i].id!);
+              showDialog(
+                context: context,
+                builder: (context) {
+                  return AlertDialog(
+                    title: const Text('Confirm Deletion'),
+                    content: const Text(
+                        'Are you sure you want to delete this scan?'),
+                    actions: <Widget>[
+                      TextButton(
+                        onPressed: () {
+                          Navigator.of(context).pop();
+                        },
+                        child: const Text('Cancel'),
+                      ),
+                      TextButton(
+                        onPressed: () {
+                          Provider.of<ScanListProvider>(context, listen: false)
+                              .deleteScanById(scans[i].id!);
+                          Navigator.of(context).pop();
+                        },
+                        child: const Text('Delete'),
+                      ),
+                    ],
+                  );
+                },
+              );
             },
             child: ListTile(
               leading: this.type == 'http'
-                  ? const Icon(Icons.home)
+                  ? const Icon(Icons.link)
                   : const Icon(Icons.map),
               title: Text(scans[i].value),
               trailing: const Icon(Icons.keyboard_arrow_right),
               onTap: () {
                 launchURL(context, scans[i]);
+              },
+              onLongPress: () {
+                Clipboard.setData(ClipboardData(text: scans[i].value));
               },
             ),
           );
